+++ foxes/variables.py
@@ -77,6 +77,7 @@ amb2var = {a: v for v, a in var2amb.items()}
 AMB_RPOINT_RESULTS = "amb_rpoint_res"
 
 K = "k"
+KB = "kb"
 KY = "ky"
 KZ = "kz"
 KTI = "kTI"